backport of commit a0176356f73f4724a44985471d29b85f3dc05b77
